🔨 refactor(SobyteAuthButton.kt): add support for scaling and rounded corners on click 🔨 refactor(SobyteLinearGradient.kt): remove unused imports 🔨 refactor(ModifierExtensions.kt): add support for corner radius in scaleOnClick function The SobyteAuthButton.kt file now supports scaling and rounded corners on click. The SobyteLinearGradient.kt file has had its unused imports removed. The ModifierExtensions.kt file now supports corner radius in the scaleOnClick function.
🎨 style(LandingScreen.kt): adjust UI elements and add login icon
💄 chore(LandingScreen.kt): change typography style and logo size
🐛 fix(LandingScreen.kt): add padding to top of content to avoid hiding behind status bar
The changes made to the LandingScreen.kt file include adjusting the size of the logo and changing the typography style. The UI elements were also adjusted to improve the overall look of the screen. A login icon was added to the login button. Padding was added to the top of the content to avoid it being hidden behind the status bar.